Store checker ID "NOT FOUND" if not found.
Some checkers may have no name. This can be caused by misinterpreting a
plist file or by the fault of a checker. In the long terms if such a
missing name is found, it should be checked individually for the given
checker. This commit is mainly for GUI purposes, because in this case a
dangling node appears in the filter field and choosing it results
selecting all findings.

This commit fixes #422.