Use an external json file for storing checker messages
The clean solution would be to place `checker_message_map.json` in the
config directory. To be able to get its path the
`generic_package_context` should be extended with a getter function.
Unfortunately unit tests don't have a context from which this could be
get. That't the reason why the `.json` file is placed in the
`codechecker_lib` directory so it can be referred relatively to the
script which reads it. By the way this `.json` file is specific for the
plist parser which needs it, so maybe it's not so ugly to handle them
together.
Fixes #450