Build command storage changed to build cmd hash
In the previous releases the whole original build command from the compilation
command json file was stored. It could take too much space in the database
if the build commands were long (no compression was used).
To address this issue in the commit ffef1ebfb1a831d84d925bef8df9dd78b8ae2ed7
storing the build command was only done in debug mode.
With that solution the update mode did not work correctly because it
depends on the build command right now.

With this commit the hashed build command from the compilation database
is stored to address the storage size problem and the update mode can
use the stored hash values.