Fixing blinking arrows.
CodeMirror renders only the currently visible lines (+-)10 lines by
default. This means that in case of a bug path arrow the source or the
target of the arrow is not rendered where one of the end of the arrow
can bind. Here we raise number of rendered lines. 500 is an empirical
value. Loading the file and rendering these lines seems to be fast
enough, and the arrows have stopped blinking too.

Fixes #518