Parsing (tidy) and storing control node fix
The tidy output parser created control nodes with even if the edges were
empty. Without range information a control section can not be used.

The result storage handler assumed that if a control node exists there
will be edges in it. If there is no edge no information can be stored
about the source positions.

Tidy output parser is modified to do not create a control sections if no
edge is available and storage handler will skip control nodes without
any edges in it.