Arrow background color fix
Fixes #650.
The plist file store the source and the target of an arrow as two
ranges: the source of the arrow is a (line-col, line-col) pair which
define and a range. The target of the arrow is a range as well. The bug
was that the coloring went from starting point of the source range to
the end pont of the target range. In this commit the two ranges are
added to the database separately.