update compilation command handling
In the current solution the update mode depends on the compilation
commands. If not given to the store command cleaning up the
previous reports will not work properly.

Introduce a new argument for the store command where the
compilation database (json) generated by the log can be used to
get the compilation commands for each source file.