fix if workspace is removed before server stop
It is possible that the workspace was removed before the
server was stopped in that case an IOError was thrown
by the instance manager during the rewriting the configuration
files.

If stopping the server takes more time in the comment tests
the cleanup fails beacause there are files used by the server.