Storage on remote server (non localhost) is fixed.
Source files are taken from the ZIP file instead of the directories on the server.
content_hashes.json file is now generated and sent over by the store command.
This file contains the content hashes of all source files that are referred to from the plist.
So the server can now find the correct file versions for the source files when storing the reports.

Fixed a bug which caused source files not being put into the ZIP file when their content hashes were the same.

Renamed necessaryFileContents to getMissingContentHashes to better express its functionality.

Detection status will be marked to new only if the bug path differs in file name (and not full path).