Improve result query performance
To get the last bug events is a costly query just to view
the line number on the UI.
The Report table is extended with the line and column data for
faster query. The api to get the reports was changed to return
the line and column data.
Report events should be queried only if they really needed.