fix skip for header files
Skipping reports in header files included by a c/cpp
file can be done only after the analysis is done and the
plist output is generated.

With this modification if there is a skip file with headers
in it, the reports from the header will be removed from the
generated plist file.