Fix a storage bug
Fixes #878
In the CTU parsing there happened to be a path which goes threw a file
twice, including another file between these two. In the previous
implementation we had a dictionary which contained the files as keys,
but according to the precondition of `collect_path_events()` function
the keys of this dictionary have to be the file list in the same order
as in the plist file including multiple occurrences.