Source language set to c++ when g++ is called. ld-logger logs the full path of the compiler (insted of only basename) so that compiler settings auto-detection can be done by the analyze even if the compiler is not in the PATH when analyze is called.
Include path auto-detection is switched off from ld-logger
and always done by the analyze sub-command.
--add-compiler-defaults flag gets deprecated as auto-detection
is always performed (even if it is not set).

Fix logger test

Since `--add-compiler-defaults` is forced, the
`--target=...` flag also becomes the part of the build command when
analyzing.

If clang is used as the compiler for the project, compilation
target is autodetected also, this way the handling becoming
identical to gcc.

build-logger got documented.

HOWTO is extended with incremental analsysis description.
We removed cross-compilation description from the HOWTO as it
is done automatically. It is not recommended anymore to
set in in the saargs and tidy-args files.