Workaround for wrong clang-tidy output
Clang-tidy has an output of the following format:

/path/to/file:<line>:<col> message [checker-name]
code_fragment(42);
^~~~~~~~~~~~~

The output may contain an optional fixit. Several groups of these lines
may follow each other.
In case of bugprone-undefined-memory-manipulation checker it happened
that in the last group only the first line appeared. The tidy output
converter assumed that the following lines are also present. If not then
it was an infinite loop.
Fixes #1064