fix logger initialization
The global LOG should be initialized only after the setup_logger
call which reads up the log config.

If the LOG is initialized at module import time it will use the
default Python logging config. By default this will print the
messages to the standard output which might not be a problem.

Some log messages might go to the wrong place (stdout) before the
logging config is read up in setup_logger if the default setting
in the log config forwards the log messages not to the stdout.