check source file modification date before store/parse/plistToHtml
If the source file modification date changes between the analysis
and the report processing it can indicate that the content of the
source files did change as well.

Changes in the source files can cause that the results will not be
shown properly after store, parse or at the html report generation.

Checking source file modification dates are done before:
* store (store will fail if there is a change)
* parse (reports with a source file change will not be printed and
         will not count into the statistics)
* plistToHtml (html files will only be generated for the reports where
               the source files did not change)