Remove the default log level
Remove the default log level from command line arguments because
it had always overridden the log level which came from the logger
configuration file.
This way if the log level is not specified in the command line
it will come from the configuration file. If the log level is
specified in the command line it will set the log level of
handlers and loggers to this.