fix Default product config modification
The sqlite database path for the Default product was set up and stored
as a relative path. Any config modification on the UI caused an error.
The sqlite db path is converter to absolute path every time a change
is requested but it did not match the relative path stored in the database.