Skip reports at store
The skip file content was not used at store.
If the same report directory is used without a skip file
and with a skip file the reports which should be skipped are
stored even if a skip file was used for the second analysis.

This change stores the skip file content in the metadata.json so
the store in the server can use this information to skip the
required reports.

Changes in the tests were required to create this use case.

This solution still sends all the reports and files to the
server and it does report skipping there.
It can be improved later where we do more fine grained file detection.
Right now the problematic files are mainly the header files
which can belong to multiple reports (paths) and if the
header file is skipped at the client side of the store
the server will not be able to show the missing file when viewing the report.