Setting up checker severities
-Specifying severities for yet uncategorized checkers.
Many checkers are alias of each other. For them, the same
severity is specified.

-alpha.cplusplus.DeleteWithNonVirtualDtor is put into default
checker package as it seems to be working with a low false positive
rate.

-optin.portability.UnixAPI is put into the default package
as it warns for errorneous basic posix function calls (malloc with 0 size).

-new core.NonnilStringConstants is put into the default package as
it models basic const char strings, which otherwise would bring in false positives.