Remove the build output from the build action hash.
Since version 6.3 the build action hash contains the build output,
which led to multiple analysis of files with essentially the
same build actions compiled into different outputs
(due to their different hash).
This is unnecessary, since the output does not change the analysis,
and results in a performance drawback. Removing the build output
from the hash seems to increase the performance without
affecting the analysis results.