Query reports only when shown.
When opening a product, all reports are queried even if "All reports"
tab is not active. In case of >200000 reports this runs a huge query in
the background which occupies a whole thread for a relatively long time.
With this modification the reports are queried only when the grid is
shown.