Zombie processes remain on analysis interruption
When the analysis stops with Ctrl-C then the subprocesses in the process
pool are not terminated. The reason is that the shared data structures
managed by multiprocess.managers.SyncManager() type die before the
termination of processes. This causes an exception in pool.terminate()
which makes the subprocesses of the pool zombies.