Fix instance manager file lock
It is possible when multiple CodeChecker servers are started in the
same time that multiple server instances are trying to write to the
same `.codechecker.instances.json` file. Some of them will not
append but override the already existing entries. This way not all
of the running server instances will be registered in the instance
file.
This commit solves this problem by locking this file properly.