Context free hash
The bug identification hash generated by Clang Static Analyzer
contains the function name in the hash calculation. This is not
appropriate for projects where the name of the functions change
frequently. This commit introduces a new argument for the analyze
and check subcommands which would calculate the hash value as
Clang Tidy. If this switch is used, the bugs will be identified
with the CodeChecker generated context free hash even though the
Clang Static Analyzer generated a context sensitive hash.