Merge and simplify log_parser and option_parser
The log_parser and option_parser modules are merged into one file. The
input of the log_parser is a compilation database as a Python list and
the output is a list of BuildAction object which contains the necessary
information for checking the given TU.

The --compiler-include-file, --compiler-target-file and
--compiler-info-file flags are emitted from the command line, because
these have no use-cases. However, the compiler implicit information are
dumped to a JSON file in the report directory for debug reasons. But
these are not read while analyzing.

The compiler flags which are necessary for the linker are also skipped
and not stored in the BuildAction object, because these are not used
anywhere.