New notification banner
Defined new Configuration endpoint for general server related
configuration.

Modified the header pane style to use flexbox, and added a new
Notification banner that displays SuperAdmin specified text.

The notification text can be specified from the product list page with a
button.

A now table was created in the configuration database. This table stores
the configuration values in a key value style.