Handling ccache in build commands
When ccache is configured on a system then it acts as a compiler: it gets the
further command line options. However, in this case we can't fetch the implicit
compiler information, like implicit include paths or the default standard. In
case the ccache command is followed by a compiler, we change the original
build command so the ccache is skipped from the beginning.