Resurrect --compiler-info-file analyze flag.
This option was removed by 5519409 under the assumption that it is not
used anywhere. The truth is that this flag is very useful for debugging
purposes and some debug tools in the /scripts directory do use it.
This commit re-introduces the --compiler-info-file option on top of the
refactored log parser.