Collect compiler info for both languages C/C++
Collect compiler includes and standards for both
language (C/C++) independent from the actual compilation
command and select the right values at analyzer command
build time.

It is possible that in a compile command json the same
compiler binary is used to build C and C++ source.
In this case the include paths and the standard will be
detected for the first command but it will be used for the
command too which is wrong.