[scripts/debug_tools] Add CTU crash reduction script.
This script uses the existing debug tools to create a debug environment,
then preprocesses all source files in the failed ZIP, and runs C-Reduce
on them. As an optimization, it empties all source files that are found
not to contribute to the crash.

Based on Péter Szécsi's original script.