use execve to call the main script
the outer script is just a wrapper to setup the env for the main
script and to save the original environment for logging.
It is not required to run after the setup and save was done.

Cleaning up the saved environment file is done not in the initializer
but in the top CodeChecker.py script when it exists.