Improve command line argument escaping
In this fix we interpret escaping from the compilation database as the original
command meant. In this commit the unnecessary "smart" escaping transformations
are eliminated, because they could result in a transformation from "hello" to
\"\"\"hello\"\"\".
In CodeChecker we are storing the analyzer options as lists which can directly
be passed to Popen.