cleanup make targets with/without virtualenv
There are two targets for each tests the default target assumes
an anlready sourced virtualenv with all the dependecies.
The target names ending *in_env will automatically setup
and source a virtualenv to run the tests.

In travis ci nose needs to be installed explicitly because
pip detects an already installed version under numpy but that binary
is not visible in the PATH.