The skipped flags are skipped in case of Clang too
The idea is that in skip-list we are enumerating the flags which
are consumed by GCC but not by Clang. Theoretically there shouldn't
be any skipped flag if the original command used Clang since these
are known only by GCC.
However, in case of Clang we'd like to skip flags starting with
-Xclang which is also listed in our skip-list. -Xclang -emit-llvm is
such a flag that makes Clang emit an LLVM IR instead of the PList file.