fix file modification time check at store
When comparing the source code modification times for each plist
only the source files included in the plist file should be checked.
The source files referred in the plist file are important during storage
of that specific plist report.

Previously in the comparison for each plist report
the file modification times from all of the collected source files were used.
Which is not necessary and can lead to storage errors if some non
related file was modified.