[makefile] Build logger libraries for multiple targets
With the current setup we only build the ldlogger.so for that
architecture where the package was built, but during runtime in
a different environment it is possible that the 32bit shared
objects are needed too.

With these changes `make package` by default will build ldlogger
shared objects for `32bit` and `64bit` too. We also introduced a
new environment variable `BUILD_LOGGER_64_BIT_ONLY` what you can
set to `YES` before the package build to build and package
`64 bit only` shared objects and ldlogger binary.