[fix] Scroll to last selected element
When we are in the Bug overview page and we click on a bug a new tab
will be opened. If we go back to the Bug overview tab it will take us
to the top of the bug list which is irritating when looking through
a long list of defects. This commit will solve this problem by
scrolling back to the last selected item on the Bug overview page.