Fix ccache handling
CCache basically has two ways to use:

ccache <compiler> <flags>
ccache <flags>

Currently we intend to support the first case. For this we have to
determine whether it is a compiler at the second position. Earlier
we checked if it matches on CC_LOGGER_GCC_LIKE. However, if this
environment variable is set to "ccache" (since the user wanted to
log ccache invocations directly) then the compiler at the second
position won't match. In this commit we consider the second parameter
a compiler if it is executable in the current environment.