TU collector crash when preprocess warning
The compiler can emit the files of a translation unit with the help
of the preprocessor. The output of the compiler call thus can be a
file list: gcc -E -M -MT __dummy main.c
However, if there is a compiler warning at the beginning of this
command's output then TU collector considered the words of the
warning message as file names.