New test for storage with file modification
Storage should be successful if a non report related file changed.
Checking the modification time stamps should be done only for
the source files mentioned in the report plist files.
Modifying any non report related file should not prevent the storage
of the reports.

This is a test for a previous fix in: 5cb63595