[server] Use DBSession object to create new session
Unfortunately we forgot to close `run_db_session` database session
in the `__get_product` function which caused idle database transactions.
To solve this problem now we are using `DBSession` object to create
new session which will be automatically closed.

Also an unused session was removed too from the `config_handler.py` and
the `DBSession` context manager was introduced for the product related
sessions so they are closed all the time.