Logger: compiler on absolute path + refactoring
The compiler name was indicated as toolName several places in the
logger. This was just the name of the binary without an absolute
path at the beginning. This toolName was also stored in the
LoggerAction objects, but it was not used anywhere, so it has been
removed.

The loggerGccParserCollectActions() and
loggerJavacParserCollectActions() functions had 3 parameters: prog_,
toolName_ and argv_. The first one is the program name caught by the
exec* commands, i.e. the compiler, usually with an absolute path that
may contain symlinks in it. The second one is the longest postfix of
prog_ which doesn't contain a "/" character. The third one is the
original argv_ array which also contains the compiler name at the 0th
position in its original form as it was called.
In this commit I removed the second parameter to decrease the
confusion among the three forms.

Earlier the location of the tool name was determined by looking for
it in the $PATH environment variable. For example: suppose that the
user emitted the following command:

/usr/bin/gcc main.c

Suppose that there is a compiler at /a/b/c/gcc in the $PATH. Since
the toolName_ is gcc, we found it by "which" algorithm that gives
/a/b/c/gcc as result, even if the user chose /usr/bin/gcc directly.

The logger converted the directory paths after -I and -L to absolute
if they were not separated from -I and -L flags with a space
character. This conversion is now eliminated.