[server] Fix default permissions
Previously some permissions were default granted if the server was
started with authentication, which means that initially, every user
has the permission. With these changes all permissions are
`default not granted`, which means that only users whom are explicitly
given the permission have it.

By default, the `start_or_get_server` function in the test cases
created a server which required authentication. With these changes
we will create 2 servers - with authentication and without it.
The `add_test_package_product` function will also give permissions
for the current product to some users.