optimize context creation in massStoreRun api call
Getting the context through webserver_context resulted
loading all of the server config files (parsing json)
in each massStoreRun api call, this extra work should not
be done in the request handling methods.
Loading the config files should be done only once
at the server start or when the reload is triggered
by an external signal.