[server] getRunData limitation
Previously `getRunData` function returned all the runs. It is possible
that the user stored a lot of runs (`>1000`) so getting these data was
too slow. For this reason we introduced a new limit/offset parameters
for the getRunData function which limits the number of runs which
will be returned to the user.

We've also introduced a new API function called `getRunCount` which
will return the number of available runs. This can be filtered by
a run filter parameter.