[server] New API function for get check command
This commit introduces a new API function to get check command
for a run or a run history. This will reduce the HTTP response
message size because we will not send back the check command to
the client when getting run data or run history data.

In the web client code we will use this function to get the check
command when the user clicks on the Show link on the run or run
history pages.