[server] Run history limitation
Use `DataGrid` and `Store` at Run history page to show the items.
This will allow us to load run history on page scroll.

Create a new API function called `getRunHistoryCount` to get the
number of run history based on a filter set. This is needed by the
`DataGrid` to support page scroll.