[server] Increase performance of getting a product
There is no need to connect to each product databases when getting
the product list because it will cause bad user experience when
loading the product listing page.

With this patch first we will try to get product details from
the product database by using sqlalchemy and if it is failed
we will try to connect to the product which will initiate the
actual connection to the database configured for the product
and update the database status then we will try to get the
product details again.