[server] Create separate API function to get analysis statistics
At the `getRunData` and `getRunHistory` functions we returned the
whole analysis statistics but we used only the number of failed and
successfully analyzed files fields. To reduce the response message
size of these API functions we created a separate API function called
`getAnalysisStatistics` to get analysis stastistics for a run which
will be called when the user clicks on the Show link on the Run List
and Run History pages.