[server] Print an info message when the server is shutting down
If we print something after the `ThreadPool` is terminated first
the server is being stopped, the terminal will be active and the
message will be printed to the active terminal. To reproduce this
problem, start a server and hit Ctrl+C.

With these changes we do not print a message after the HTTP server
is terminated.