Skipping -specs flag in TU collector
This flag can be given a .specs file which contains the config options of
cc1, cc1plus, as, ld, etc. Sometimes this file is just a temporary during
the compilation. However, if the file doesn't exist, this flag fails the
compilation. Since this flag is not necessary for dependency generation,
we can skip it.