Collect compiler information for multiple languages
If the same compiler was used multiple times in a
compile command json but compiled multiple source
files for different languages (C and C++) the implicit
compiler information was collected only when first detected
(only for C).

With this change when a compiler is first found the include
paths, default build targets and compiler standards
will be detected for both languages. So the information can
be used when the same compiler is compiling for a
different language.

Additional changes were done to set the include paths,
build targets and compiler standard at multiple analysis
stages like, statistics collection, ctu pre analysis and
for setting the clang tidy arguments.

Debug tools were updated to produce the new compiler_info format
based on the limited information in the previous compiler info files.

In case of the different ccache modes we depend on the correct compiler
detection (in the earlier steps) to get the implicit compiler
information through ccache.