fix plist parsing for skip files
The introduced new plist xml parsing method with lxml
was not introduced for the use cases where a skip file
was used. In case of a skip file the report plist files
can be rewritten (the reports based on the skip file content)
are removed.

Previous tests did not fail because a second skip is done
at the server side during storage and the server skip tests
worked on the stored reports.