[cli] Gettings results for runs which contain colons
Filtering and getting the results for run names with colon
punctuation mark does not work properly. If there is a timestamp
in the run name we want to filter "myrun-12:23:12" the punctuation
mark will be used as a delimiter of multiple run names even if
this is not the case.
For this reason we set the `nargs` option to `+` for the run names
option of the `CodeChecker cmd results` subcommand so we will be
able to get results for run names which contains colon punctuation
marks.

E.g: `CodeChecker cmd results myrun-12:23:12 myrun-22:00:10`