exit on missing checker or profile names
It is possible that there is typo in the checker name
or the checker was removed/renamed in a newer clang version.
The analysis starts but the user will not get any results
because there is no checker with that name.
To prevent this exit before the analysis if a checker name is missing.

Every checker argument is checked because it is possible
that multiple checkers are missing or there are multiple typos.