ctu display progress flag changed in clang8
The ctu-display-progress flag did change in clang8.
The argument is turned on in case of ctu failure to
collect the source files used by ctu for analysis.

Analyzer config options can be detected only by listing
all the options and search for the required one.
The previous solution for analyzer options (to
compile a file) will not work for analyzer
config options (clang returns 0 even if there is no
such analyzer config option)

Not just the flag but the analyzer output format did
changed in clang8 too.