[server] Refactoring review status change permission check
Previously in the `_setReviewStatus` function we checked the
required permissions. The problem was here that this function
was called multiple times in the `__store_reports` function
which was called by the `massStoreRun` function but this was
unnecessary because we already checked the required permissions
in the `massStoreRun` function.

The `_setReviewStatus` function is also called in the
`changeReviewStatus` so now we check the required permissions in
the begining of this function too.