[web] Refactoring source code comment handler
`SourceCodeCommentHandler` object was created multiple times when
iterating over reports because we had to pass the source file
path to the constructor. With this patch we refactored the
`SourceCodeCommentHandler` class to pass the source file name
for each function which are require it so there is no need to
create new object so many times.