[web] Send plist in json format to the server
We are parsing plist files on the client and the server side when
storing analyzer results. The performance of parsing a plist file
is really bad so if the user analyzed a large project which generated
a lot of plist files parsing these files will take a long time.

With this commit we will reduce the overhead of the servers by
creating json files from the plist file and sending these files to
the server. Loading a json file will be much better then loading
and parsing a plist file.