Distinguish reports of different checkers
When a checker is alias of another then it is possible that two
checkers are reporting the same message. We wouldn't like to unique
them because it was random which checker won the possibility of
reporting. Now both checkers are emitting the report and the user can
explicitly choose to switch off any of them.

This commit also contains some improvement on checking "skipping". The
operator "|=" is not lazy and it is not always necessary to evaluate
the right hand side.