refactor suppress file handler
Using a suppress file as a store for review comments by the server
was removed a long time ago. The corresponding api call
was kept to keep backward compatibility but it is deprecated
and not used by the recent clients. Unused code was removed.

Functional tests are split to tests the related funcionality.
 - the analyze part has the feature to export suppress/review comments
   to a file
 - the server has the functionality to import suppress/review comments
   from a file