return 0 in case no analysis is needed
If the compile command json is empty CodeChecker
returned with 1 which indicates and error which is might
not the case.
The compiler command json can be empty for several reasons:
- logging the compile commands failed for some reason
- no compilation actions were called by the buid system (noting changed)

The tests needed some modification because the output
is different for calling log, analyze and parse if the compile command
json is empty (parse is called in the tests even there is nothing to
parse), compared to calling check which will exit after the analyze
part and not call parse.
With the modified tests there is some additional file which is
not skipped so there are reports to parse.