[fix] Fix getDiffResultsHash API function
Let's suppose that a report can be found in the database in a run with
detection status of `Resolved` and this report reappeared again in the
report directory. If we do comparison between this report directory
and the remote run the newly introduced report will not be shown as
new report because it is already can be found in the database.

This commit will solve this problem by filtering out reports which
have detection status of `Resolved`, `Off` or `Unavailable` when
getting diff report hashes for `NEW` and `UNRESOLVED` diff types.

This commit also changes the `diff_local_remote` test cases by storing
an other analyzer report directory to the remote run so we will have
reports which will have detection status of `Off`.