[server] Fix login form autocompletion
In Firefox the login form autocompletion didn't work. The problem
was that the `autocomplete` attribute of the `username` input field
was set to `off`. In this commit we set it to `on`.

We are also using the `get` function on the input fields to get the
values because the value field wasn't set properly in Firefox.