[server] Fix comment update
When we opened a bug which had a comment and we clicked on the edit button,
changed the comment message and pressed the `Save` button nothing happened.
The problem was that we tried to call the `replace` function on the dojo
`ContentPane` object not on the content itself. With this commit we fix
this problem.