[server] Fix local session creation
We were not able to edit the announcement header because the
`is_root` field had not set properly. The reason why it didn't
work was that if the user doesn't have a valid session token in
the database we will try to create a local session but we do not
check that the user has an entry in the SystemPermission table.
With this patch we will fix this problem.