[server] Handle exceptions in do_POST
Is it possible that a Python exception was raised in a Thrift API
function. We will catch this exception in the do_POST function and
we will try to send back a valid thrift message to the client. First
we try to call the `readMessageBegin` function on the input protocol
(iprot) but it will wait indefinitely because we have called this
function in the `processor.process(iprot, oprot)` function already
and no newer message can be found in the buffer.

To solve this problem we will create a new JSON protocol (`iprot`) by
using the value from the previous one and send a thrift exception
by using this new input protocol.