[analyzer] Add version support for profile to checkers mapping
Add the ability to specify a version-string regex which specifies which
configuration would be used to map profiles to checker-strings.
Check and warn if no checkers are enabled because of there is no profile match
for a given analyzer version.
Do not run the analysis for an analyzer if there are no checks enabled for it.
Add new version parser for clang-tidy analyzer.