keep all flags if the analyzer is the same as the compiler
If clang compiler is used to build and to analyze a project
do not filter the compiler options, keep them all.

clang and clang++ are handled as they were the same binaries.

if the compiler is clang and the analyzer is clang too only a minimal
compiler option filtering is done.

- source, output and other files are removed from the compile command
- include paths are converted to absolute path required for the reports
- preprocessor flags are filtered out (the same way as for gcc/g++)

__collect_clang_compile_opts was introduced to
collect all the compile options should be run at
last everything should be filtered out earlier

__collect_compile_opts was split up
so the include path modification and the option filtering
is done separately, because the include path modification
is required for clang but the filtering not

handling '--sysroot=' flag format needed to be changed

--sysroot flag format can be --sysroot=/path/to/include
In this case before the normalization the '=' sign must be removed.
We put back the original
--sysroot=/path/to/include as
--sysroot /path/to/include
which is a valid format too.

New test file was added to see if the version parser does not detect
gcc as a clang compiler.

compiler implicit info collection is not done for clang.