[Statistics] Replace analyzer option `api-metadata-path` with package option `APIMetadataPath`
Clang 9 enables package options beside analyzer and checker options.
This allows us to use package option instead of analyzer option for
passing the path of the YAML files needed by the statistics-based
checkers. The advantage of this approach is that the plugin now works
with the official Clang 9 release, no patching is needed. Since the
statistics-based checkers are in the same package, the option needs only
to be passed once for the whole analysis.

This patch adjust CodeChecker to work correctly with the new plugin.